// src/discordPresence/types/stashPlugin.d.ts
interface IScenePlayerProps extends React.PropsWithChildren {
  scene: Scene;
  hideScrubberOverride: boolean;
  autoplay?: boolean;
  permitLoop?: boolean;
  initialTimestamp: number;
  sendSetTimestamp: (setTimestamp: (value: number) => void) => void;
  onComplete: () => void;
  onNext: () => void;
  onPrevious: () => void;
}